Rename: 컨트롤러 폴더와 레포지토리 폴더 분리
